+++ src/idc/filter/__init__.py
@@ -17,6 +17,7 @@ from ._randomize_records import RandomizeRecords
 from ._record_window import RecordWindow
 from ._remove_classes import RemoveClasses
 from ._rename import Rename, RENAME_PLACEHOLDERS, RENAME_PH_NAME, RENAME_PH_COUNT, RENAME_PH_PDIR, RENAME_PH_SAME, RENAME_PH_PDIR_SUFFIX, RENAME_PH_EXT, RENAME_PH_OCCURRENCES, RENAME_PH_HELP
+from ._rgb_to_grayscale import RGBToGrayscale
 from ._sample import Sample
 from ._split_records import SplitRecords
 from ._strip_annotations import StripAnnotations